profile screen draft done, need to fill in modals
